pyproject.toml
[project]
name = "day1-basic-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit>=1.0.17",
    "livekit-agents>=1.2.17",
    "livekit-plugins-deepgram>=1.2.17",
    "livekit-plugins-openai>=1.2.17",
    "livekit-plugins-silero>=1.2.17",
    "python-dotenv>=1.2.1",
]
